+++ RELEASE_NOTES.md
@@ -5,7 +5,64 @@
 FIXES:
 
 1. Fix typo in Quectel PAIR650 GET message definition.
-1. Allow string type (*as well as datetime/time type*) for NMEA DT, DTL, DM and TM attribute types. TM strings must be in format "hhmmss" (or "hh:mm:ss") and DT/DTL/DM strings must be "yyyymmdd" (or "yyyy-mm-dd").
+1. Allow string type (*as well as datetime.date/time type*) for NMEA DT, DTL, DM and TM attribute constructors. TM strings must be in format "hhmmss" (or "hh:mm:ss"). DT/DTL/DM strings must be "yyyymmdd" (or "yyyy-mm-dd"). See examples below:
+
+```python
+from datetime import datetime
+
+from pynmeagps import SET, NMEAMessage
+
+# NOTE THAT LAD/NS ("N"/"S") and LND/EW ("E"/"W") attributes do not need to be explicitly
+# provided - these values will be derived from the sign of the decimal lat/lon values.
+
+# NMEA Date (DM, DT, DTL) and Time (TM) attributes can be populated in any of the following ways:
+
+# A) use formatted string types for TM and DT attributes
+msg1 = NMEAMessage(
+    "P",
+    "GRMI",
+    SET,
+    lat=-115.81513,
+    lon=37.23345,
+    date="2025-09-12",  # "-" delimiters are optional
+    time="12:15:34",  # ":" delimiters are optional
+    rcvr_cmd="D",
+)
+# <NMEA(PGRMI, lat=-115.81513, NS=S, lon=37.23345, EW=E, date=2025-09-12, time=12:15:34, rcvr_cmd=D)>
+# b'$PGRMI,11548.90780,S,03714.00700,E,120925,121534,D*3B\r\n'
+print(msg1)
+print(msg1.serialize())
+
+# B) use datetime.date() and datetime.time() types for DT and TM attributes
+msg2 = NMEAMessage(
+    "P",
+    "GRMI",
+    SET,
+    lat=-115.81513,
+    lon=37.23345,
+    date=datetime(2025, 9, 12).date(),
+    time=datetime(2025, 9, 12, 12, 15, 34).time(),
+    rcvr_cmd="D",
+)
+# <NMEA(PGRMI, lat=-115.81513, NS=S, lon=37.23345, EW=E, date=2025-09-12, time=12:15:34, rcvr_cmd=D)>
+# b'$PGRMI,11548.90780,S,03714.00700,E,120925,121534.00,D*15\r\n'
+print(msg2)
+print(msg2.serialize())
+
+# C) use default values
+msg3 = NMEAMessage(
+    "P",
+    "GRMI",
+    SET,
+    lat=-115.81513,
+    lon=37.23345,
+    rcvr_cmd="D",
+)
+# <NMEA(PGRMI, lat=-115.81513, NS=S, lon=37.23345, EW=E, date=2025-09-18, time=14:37:20.373212, rcvr_cmd=D)>
+# b'$PGRMI,11548.90780,S,03714.00700,E,180925,143720.37,D*18\r\n'
+print(msg3)
+print(msg3.serialize())
+```
 
 ### RELEASE 1.0.52